Added ssl sockets for server-server communication.
SSL sockets are used to 1. keep data confidential between servers,
but also to verify that the webserver and actuator server are who
they say they are. Included are some scripts for generating self-signed
certificates.

These self-signed certificates should be replaced with actual
certificates, and some secondary check should be put in place
(make sure that only connections from IP X are accepted).

We should consider using a service like Let's Encrypt for these
certificates. (Automated and free.)

Also in progress is communication between the python code and the
client. The hope is to lower latency by connecting directly to the
python stuff, instead of using node.js as a sort of relay. This should
also help when we get around to webcam stuff.