Created endpoint to request timeslot.
Issues: This might be a display issue,
could also be an issue with adding to the DB, but requests seem to take
an extra slot after them.

Also, I can still select slots
I already requested, which is an issue.

Another issue is that the xhr request
calls the onreadystate
function with readystate = 4 about
4 times, it seems, when it should be once.