Switched to new logger
I didn't think of it until I was almost done but it would really help to
have it print the file the messages are comming from because it doesn't do that
automatically. I'm gonna see if there's a way built in or if not I'll go back
through and do it manually.